\name{cvrt}
\alias{cvrt}
\docType{data}
\title{
	Artificial data for CorrMeta sample code: Covariates.
}
\description{
	Artificial data set with 7 covariates across 50 samples (20 MZ twins, 20 DZ twins, 10 single samples). 
	Columns of the file match to those of the expression and genotype data sets.
}
\format{
	\tabular{llll}{
		id 			\tab sample0001   \tab sample0002 \tab \ldots \cr
		cvrt000001	\tab 0.8477        \tab 0.2563      \tab \ldots \cr
		cvrt000001  \tab 0.3443        \tab 0.8324    \tab \ldots \cr
		\ldots		\tab \ldots		\tab \ldots		\tab  \ldots \cr
	}
}
\references{
	The package website: \url{http://www.bios.unc.edu/research/genomic_software/CorrMeta/}
}
\seealso{
	See \code{\link{CorrMeta_testAll}} for reference and sample code.
}
\keyword{datasets}
