\name{snps}
\alias{snps}
\docType{data}
\title{
	Artificial data for CorrMeta sample code: Genotype.
}
\description{
	Artificial data set with 7 covariates across 50 samples (20 MZ twins, 20 DZ twins, 10 single samples). 
	Columns of the file match to those of the expression and covariate data sets.
}
\format{
	\tabular{llll}{
		id 			\tab sample0001   \tab sample0002 \tab \ldots \cr
		snp000001	\tab 0        \tab 0      \tab \ldots \cr
		snp000001  \tab 1        \tab 1    \tab \ldots \cr
		\ldots		\tab \ldots		\tab \ldots		\tab  \ldots \cr
	}
}
\references{
	The package website: \url{http://www.bios.unc.edu/research/genomic_software/CorrMeta/}
}
\seealso{
	See \code{\link{CorrMeta_testAll}} for reference and sample code.
}
\keyword{datasets}
